Make revenue a long

+++ DustyPig.TMDB/Models/Movies/Details.cs
@@ -55,7 +55,7 @@ public class Details
     public DateOnly? ReleaseDate { get; set; }
 
     [JsonPropertyName("revenue")]
-    public int Revenue { get; set; }
+    public long Revenue { get; set; }
 
     [JsonPropertyName("runtime")]
     public int Runtime { get; set; }
